requirements.txt
blinker==1.7.0
boto3==1.28.83
botocore==1.31.83
Brotli==1.0.9
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
dash==2.3.1
dash-bootstrap-components==1.1.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
ediblepickle==1.1.3
Flask==2.1.1
Flask-Compress==1.11
fonttools==4.44.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.3
numpy==1.22.3
packaging==23.2
pandas==1.4.2
Pillow==9.1.0
plotly==5.18.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.27.1
s3transfer==0.7.0
scikit-learn==1.0.2
scipy==1.10.1
six==1.16.0
tenacity==8.2.3
threadpoolctl==3.2.0
urllib3==1.26.18
Werkzeug==2.1.1
whitenoise==6.6.0
zipp==3.17.0
